#13 cambio de tipo de dato para tasas
Aunque sea un valor definido como de 1 a 1000, se considera un valor porcentual así que se debe representar como un número entero

+++ pkg/segurasenior/poblacion.go
@@ -5,13 +5,13 @@ import "errors"
 type Poblacion struct {
 	Nombre_pueblo       string
 	Poblacion_total     int
-	Tasa_natalidad      int
-	Tasa_envejecimiento int
-	Tasa_mortalidad     int
+	Tasa_natalidad      float32
+	Tasa_envejecimiento float32
+	Tasa_mortalidad     float32
 }
 
-func NewPoblacion(nom string, poblacion int, tasan int, tasae int, tasam int) (*Poblacion, error) {
-	if (tasan < 0) || (tasan > 1000) || (tasae < 0) || (tasae > 1000) || (tasam < 0) || (tasam > 1000) {
+func NewPoblacion(nom string, poblacion int, tasan float32, tasae float32, tasam float32) (*Poblacion, error) {
+	if (tasan < 0.0) || (tasan > 1000.0) || (tasae < 0.0) || (tasae > 1000.0) || (tasam < 0.0) || (tasam > 1000.0) {
 		return nil, errors.New("valor de tasas erróneo, debe estar comprendido entre 0 y 1000")
 	}
 	return &Poblacion{
